feat: move DEFAULT_BASES to the sushi package
