+++ packages/sushi/src/config/index.ts
@@ -4,6 +4,7 @@ export * from './api.js'
 export * from './bases-to-check-trades-against.js'
 export * from './bentobox.js'
 export * from './custom-bases.js'
+export * from './default-bases.js'
 export * from './default-quote.js'
 export * from './furo.js'
 export * from './default-slippage.js'
